requirements.txt
aws-cdk.core
aws-cdk.aws-ssm
aws-cdk.aws-ec2
aws-cdk.aws_iam
aws-cdk.aws_ecs
aws-cdk.aws_ecr
aws_cdk.aws_codecommit
aws_cdk.aws_codebuild
aws_cdk.aws_iam
aws_cdk.aws_elasticloadbalancingv2
aws_cdk.aws_codedeploy
aws_cdk.aws_lambda
aws_cdk.aws_cloudwatch
aws_cdk.aws_logs
aws_cdk.aws_codepipeline
aws_cdk.aws_s3
aws_cdk.aws_codepipeline_actions
aws_cdk.aws_s3_deployment
aws_cdk.aws_wafv2
aws_cdk.aws_route53_targets
aws_cdk.aws_route53
aws_cdk.aws-cloudfront